finished cogs; renamed bot files
